cleanup tracking of logs
